change-data-capture/requirements.txt
mysql-replication
redis
cryptography